Makefile: add portable mode
Add a 'portable' mode that packages all relevant flat files and helper
scripts into a tarball named 'kut-portable.tar.gz'.

This mode is useful for compiling tests on one machine and running them
on another without needing to clone the entire repository. It allows
the runner scripts and unit test configurations to remain local to the
machine under test.

Signed-off-by: Jon Kohler <jon@nutanix.com>