x86/vmx: conditionally include EPT ignored bit 10 based on MBEC support
Update the ept_access_test_ignored_bits test only ignore bit 10 if
MBEC is not supported, as the bit is not ignored when MBEC is enabled.

When MBEC is enabled, test ept_allowed for OP_EXEC_USER in ept_ignored_bit.

Update unittest.cfg to break off MBEC into a separate test suite, as using
-cpu max will automagically pick up MBEC, and it should be testable with
MBEC enabled and disabled.

Signed-off-by: Jon Kohler <jon@nutanix.com>