x86/vmx: update EPT access tests for MBEC support (needs help)
Extend ept_access_test_read_execute to cover MBEC EPT r-x case, with
the caveat that two of the cases do not currently work and are now
commented out.

Need a hand with sanity checking this, as both of the commented out
test cases produce a tight EPT violation loop on the kernel side, and
I'm unsure as of yet if its a test side issue (setup?) or what.

Tests pass with both -vmx-mbec and +vmx-mbec (for the case that isn't
commented out)

Signed-off-by: Jon Kohler <jon@nutanix.com>