x86/vmx: update EPT supervisor execute only tests for MBEC support (needs help)
Extend ept_access_test_execute_only to cover MBEC EPT --x case, with
the caveat that it doesn't actually work as expected.

Need a hand with sanity checking this, as both of the commented out
test cases produce a tight EPT violation loop on the kernel side, and
I'm unsure as of yet if its a test side issue (setup?) or what.

Tests pass with both -vmx-mbec and +vmx-mbec (for the case that isn't
commented out)

Signed-off-by: Jon Kohler <jon@nutanix.com>